Adds an object observer that raises an error when encountering the same object twice
This functionality is motivated by the potential for circular object
graphs when traversing the nested object relationships (defined using
struct in define_index_fields) during document generation. The observer
records the encountered objects and raises an error if the same object
is encountered twice, thus avoiding the SystemStackError caused by an
endless loop in the circular object graph.