Adds default inheritance behavior for Indexable classes.
Adds a default behavior for inherited classes to get the parents
indexable behavior by default. This means that if class B is
a subclass of class A and class a has some definitions, all
the definitions of class A are included to class B as well.

This is an assumption that given an indexable object of class A
we also want class B which is an subclass of A
to be indexed as well with the same index definitions of its
parent. In other words both classes should have the same schema
definition.